Refactor Zilean scraper for MediaItem support (#516)
The Zilean scraper has been refactored to support the MediaItem class. Previously, it only supported the Movie class. The 'aired_at' attribute of a Movie instance is now replaced with the 'year' attribute of a MediaItem instance. This change allows more flexibility in handling different types of media items beyond just movies.